requirements.txt
Django==1.8.7
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2

django-mptt==0.7.4
django-mptt-admin==0.2.1
six==1.10.0
django-autoslug==1.9.3
pytils==0.3
django-flat-theme==1.1.3
django-registration-redux==1.2
django-crispy-forms==1.5.2
